+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - [Feature: add rate limit filter debug log](https://github.com/Tencent/spring-cloud-tencent/pull/437)
 - [Add configurable heartbeat interval support](https://github.com/Tencent/spring-cloud-tencent/pull/444)
 - [feat:enhance Feign and RestTemplate and support Polaris monitor.](https://github.com/Tencent/spring-cloud-tencent/pull/447)
+- [Feature: support spring cloud config data.](https://github.com/Tencent/spring-cloud-tencent/pull/451)
 - [Optimize feign & rest-template circuit-breaker logic](https://github.com/Tencent/spring-cloud-tencent/pull/454)
 - [Feature: Add disposable metadata transfer support](https://github.com/Tencent/spring-cloud-tencent/pull/459)
 - [docs:update mvnw.](https://github.com/Tencent/spring-cloud-tencent/pull/476)